+++ src/main/java/com/picimako/mockitools/inspection/consecutive/SimplifyConsecutiveReturnCallsInspection.java
@@ -48,7 +48,7 @@ public class SimplifyConsecutiveReturnCallsInspection extends SimplifyConsecutiv
     }
 
     @Override
-    protected void register(ConsecutiveCallRegistrarDescriptor descriptor, @NotNull ProblemsHolder holder) {
-        doRegister(descriptor, holder, new MergeConsecutiveStubbingCallsQuickFix(descriptor.initCallsAsPointers(), TypeConversionMethod.NO_CONVERSION));
+    protected void register(ConsecutiveCallRegistrarContext registrar, @NotNull ProblemsHolder holder) {
+        doRegister(registrar, holder, new MergeConsecutiveStubbingCallsQuickFix(new ConsecutiveCallQuickFixContext(registrar), TypeConversionMethod.NO_CONVERSION));
     }
 }