+++ src/test/java/com/picimako/mockitools/inspection/consecutive/SimplifyConsecutiveReturnCallsInspectionTest.java
@@ -26,30 +26,43 @@ public class SimplifyConsecutiveReturnCallsInspectionTest extends MockitoolsV4In
             "           .thenCallRealMethod()\n" +
             "           .thenReturn(2)\n" +
             "           .then<caret>Return(3);",
-        "       Mockito.when(mockObject.didSomething()).thenReturn(1).thenCallRealMethod().thenReturn(2, 3);",
+        "       Mockito.when(mockObject.didSomething()).thenReturn(1)\n" +
+            "           .thenCallRealMethod()\n" +
+            "           .thenReturn(2, 3);",
         "       Mockito.when(mockObject.didSomething()).thenReturn(1)\n" +
             "           .thenR<caret>eturn(2)\n" +
             "           .thenCallRealMethod()\n" +
             "           .thenReturn(3);",
-        "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2).thenCallRealMethod().thenReturn(3);",
+        "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2)\n" +
+            "           .thenCallRealMethod()\n" +
+            "           .thenReturn(3);",
         "       Mockito.when(mockObject.didSomething()).thenReturn(1)\n" +
             "           .thenR<caret>eturn(2)\n" +
             "           .thenCallRealMethod()\n" +
             "           .thenReturn(3)\n" +
             "           .thenReturn(4);",
-        "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2).thenCallRealMethod().thenReturn(3).thenReturn(4);",
+        "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2)\n" +
+            "           .thenCallRealMethod()\n" +
+            "           .thenReturn(3)\n" +
+            "           .thenReturn(4);",
         "       Mockito.when(mockObject.didSomething()).thenReturn(1)\n" +
             "           .thenReturn(2)\n" +
             "           .thenCallRealMethod()\n" +
             "           .thenReturn(3)\n" +
             "           .thenRet<caret>urn(4);",
-        "       Mockito.when(mockObject.didSomething()).thenReturn(1).thenReturn(2).thenCallRealMethod().thenReturn(3, 4);",
+        "       Mockito.when(mockObject.didSomething()).thenReturn(1)\n" +
+            "           .thenReturn(2)\n" +
+            "           .thenCallRealMethod()\n" +
+            "           .thenReturn(3, 4);",
         "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2, 3)\n" +
             "           .thenReturn(4)\n" +
             "           .thenCallRealMethod()\n" +
             "           .thenReturn(5)\n" +
             "           .thenR<caret>eturn(6, 7);",
-        "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2, 3).thenReturn(4).thenCallRealMethod().thenReturn(5, 6, 7);"
+        "       Mockito.when(mockObject.didSomething()).thenReturn(1, 2, 3)\n" +
+            "           .thenReturn(4)\n" +
+            "           .thenCallRealMethod()\n" +
+            "           .thenReturn(5, 6, 7);"
     );
 
     private static final Map<String, String> DO_RETURN_WHEN_CASES = Map.of(
@@ -60,34 +73,52 @@ public class SimplifyConsecutiveReturnCallsInspectionTest extends MockitoolsV4In
             "           .doReturn(2)\n" +
             "           .doRet<caret>urn(3)\n" +
             "           .when(mockObject).didSomething();",
-        "       Mockito.doReturn(1).doCallRealMethod().doReturn(2, 3).when(mockObject).didSomething();",
+        "       Mockito.doReturn(1)\n" +
+            "           .doCallRealMethod()\n" +
+            "           .doReturn(2, 3)\n" +
+            "           .when(mockObject).didSomething();",
         "       Mockito.doReturn(1)\n" +
             "           .doR<caret>eturn(2)\n" +
             "           .doCallRealMethod()\n" +
             "           .doReturn(3)\n" +
             "           .when(mockObject).didSomething();",
-        "       Mockito.doReturn(1, 2).doCallRealMethod().doReturn(3).when(mockObject).didSomething();",
+        "       Mockito.doReturn(1, 2)\n" +
+            "           .doCallRealMethod()\n" +
+            "           .doReturn(3)\n" +
+            "           .when(mockObject).didSomething();",
         "       Mockito.doReturn(1)\n" +
             "           .doRet<caret>urn(2)\n" +
             "           .doCallRealMethod()\n" +
             "           .doReturn(3)\n" +
             "           .doReturn(4)\n" +
             "           .when(mockObject).didSomething();",
-        "       Mockito.doReturn(1, 2).doCallRealMethod().doReturn(3).doReturn(4).when(mockObject).didSomething();",
+        "       Mockito.doReturn(1, 2)\n" +
+            "           .doCallRealMethod()\n" +
+            "           .doReturn(3)\n" +
+            "           .doReturn(4)\n" +
+            "           .when(mockObject).didSomething();",
         "       Mockito.doReturn(1)\n" +
             "           .doReturn(2)\n" +
             "           .doCallRealMethod()\n" +
             "           .doReturn(3)\n" +
             "           .doRe<caret>turn(4)\n" +
             "           .when(mockObject).didSomething();",
-        "       Mockito.doReturn(1).doReturn(2).doCallRealMethod().doReturn(3, 4).when(mockObject).didSomething();",
+        "       Mockito.doReturn(1)\n" +
+            "           .doReturn(2)\n" +
+            "           .doCallRealMethod()\n" +
+            "           .doReturn(3, 4)\n" +
+            "           .when(mockObject).didSomething();",
         "       Mockito.doReturn(1, 2, 3)\n" +
             "           .doReturn(4)\n" +
             "           .doCallRealMethod()\n" +
             "           .doReturn(5)\n" +
             "           .doRe<caret>turn(6, 7)\n" +
             "           .when(mockObject).didSomething();",
-        "       Mockito.doReturn(1, 2, 3).doReturn(4).doCallRealMethod().doReturn(5, 6, 7).when(mockObject).didSomething();"
+        "       Mockito.doReturn(1, 2, 3)\n" +
+            "           .doReturn(4)\n" +
+            "           .doCallRealMethod()\n" +
+            "           .doReturn(5, 6, 7)\n" +
+            "           .when(mockObject).didSomething();"
     );
 
     private static final Map<String, String> GIVEN_WILL_RETURN_CASES = Map.of(
@@ -97,30 +128,43 @@ public class SimplifyConsecutiveReturnCallsInspectionTest extends MockitoolsV4In
             "           .willCallRealMethod()\n" +
             "           .willReturn(2)\n" +
             "           .willR<caret>eturn(3);",
-        "       BDDMockito.given(mockObject.didSomething()).willReturn(1).willCallRealMethod().willReturn(2, 3);",
+        "       BDDMockito.given(mockObject.didSomething()).willReturn(1)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(2, 3);",
         "       BDDMockito.given(mockObject.didSomething()).willReturn(1)\n" +
             "           .willRet<caret>urn(2)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(3);",
-        "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2).willCallRealMethod().willReturn(3);",
+        "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(3);",
         "       BDDMockito.given(mockObject.didSomething()).willReturn(1)\n" +
             "           .willRe<caret>turn(2)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(3)\n" +
             "           .willReturn(4);",
-        "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2).willCallRealMethod().willReturn(3).willReturn(4);",
+        "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(3)\n" +
+            "           .willReturn(4);",
         "       BDDMockito.given(mockObject.didSomething()).willReturn(1)\n" +
             "           .willReturn(2)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(3)\n" +
             "           .willRe<caret>turn(4);",
-        "       BDDMockito.given(mockObject.didSomething()).willReturn(1).willReturn(2).willCallRealMethod().willReturn(3, 4);",
+        "       BDDMockito.given(mockObject.didSomething()).willReturn(1)\n" +
+            "           .willReturn(2)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(3, 4);",
         "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2, 3)\n" +
             "           .willRe<caret>turn(4)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(5)\n" +
             "           .willReturn(6, 7);",
-        "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2, 3, 4).willCallRealMethod().willReturn(5).willReturn(6, 7);"
+        "       BDDMockito.given(mockObject.didSomething()).willReturn(1, 2, 3, 4)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(5)\n" +
+            "           .willReturn(6, 7);"
     );
 
     private static final Map<String, String> WILL_RETURN_GIVEN_CASES = Map.of(
@@ -130,30 +174,43 @@ public class SimplifyConsecutiveReturnCallsInspectionTest extends MockitoolsV4In
             "           .willCallRealMethod()\n" +
             "           .willReturn(2)\n" +
             "           .willR<caret>eturn(3).given(mockObject).didSomething();",
-        "       BDDMockito.willReturn(1).willCallRealMethod().willReturn(2, 3).given(mockObject).didSomething();",
+        "       BDDMockito.willReturn(1)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(2, 3).given(mockObject).didSomething();",
         "       BDDMockito.willReturn(1)\n" +
             "           .willRet<caret>urn(2)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(3).given(mockObject).didSomething();",
-        "       BDDMockito.willReturn(1, 2).willCallRealMethod().willReturn(3).given(mockObject).didSomething();",
+        "       BDDMockito.willReturn(1, 2)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(3).given(mockObject).didSomething();",
         "       BDDMockito.willReturn(1)\n" +
             "           .willR<caret>eturn(2)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(3)\n" +
             "           .willReturn(4).given(mockObject).didSomething();",
-        "       BDDMockito.willReturn(1, 2).willCallRealMethod().willReturn(3).willReturn(4).given(mockObject).didSomething();",
+        "       BDDMockito.willReturn(1, 2)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(3)\n" +
+            "           .willReturn(4).given(mockObject).didSomething();",
         "       BDDMockito.willReturn(1)\n" +
             "           .willReturn(2)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(3)\n" +
             "           .willRe<caret>turn(4).given(mockObject).didSomething();",
-        "       BDDMockito.willReturn(1).willReturn(2).willCallRealMethod().willReturn(3, 4).given(mockObject).didSomething();",
+        "       BDDMockito.willReturn(1)\n" +
+            "           .willReturn(2)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(3, 4).given(mockObject).didSomething();",
         "       BDDMockito.willReturn(1, 2, 3)\n" +
             "           .willRe<caret>turn(4)\n" +
             "           .willCallRealMethod()\n" +
             "           .willReturn(5)\n" +
             "           .willReturn(6, 7).given(mockObject).didSomething();",
-        "       BDDMockito.willReturn(1, 2, 3, 4).willCallRealMethod().willReturn(5).willReturn(6, 7).given(mockObject).didSomething();"
+        "       BDDMockito.willReturn(1, 2, 3, 4)\n" +
+            "           .willCallRealMethod()\n" +
+            "           .willReturn(5)\n" +
+            "           .willReturn(6, 7).given(mockObject).didSomething();"
     );
 
     @Override